update docs and build scripts

+++ scripts/build_mac.sh
@@ -3,13 +3,12 @@
 echo "========= OnTheSpot macOS Build Script =========="
 
 
-echo " => Cleaning up previous builds!"
-rm -rf dist
-
-
-echo " => Creating and activating virtual environment..."
+echo " => Cleaning up previous builds and preparing the environment..."
+rm -f ./dist/OnTheSpot.tar.gz
+mkdir build
+mkdir dist
 python3 -m venv venv
-source venv/bin/activate
+source ./venv/bin/activate
 
 
 echo " => Upgrading pip and installing necessary dependencies..."
@@ -17,17 +16,30 @@ venv/bin/pip install --upgrade pip wheel pyinstaller
 venv/bin/pip install -r requirements.txt
 
 
+echo " => Build FFMPEG (Optional)"
+FFBIN="--add-binary=dist/ffmpeg:onthespot/bin/ffmpeg"
+if ! [ -f "dist/ffmpeg" ]; then
+    curl -L -o build/ffmpeg.zip https://evermeet.cx/ffmpeg/ffmpeg-7.1.zip
+    unzip build/ffmpeg.zip -d dist
+#    cd build
+#    curl https://ffmpeg.org/releases/ffmpeg-7.1.1.tar.xz -o ffmpeg.tar.xz
+#    tar xf ffmpeg.tar.xz
+#    cd ffmpeg-*
+#    ./configure --enable-small --disable-ffplay --disable-ffprobe --disable-doc --disable-htmlpages --disable-manpages --disable-podpages --disable-txtpages
+#    make
+#    cp ffmpeg ../../dist
+#    cd ../..
+fi
+
+
 echo " => Running PyInstaller to create .app package..."
-mkdir build
-curl -L -o build/ffmpeg.zip https://evermeet.cx/ffmpeg/ffmpeg-7.1.zip
-unzip build/ffmpeg.zip -d build
 pyinstaller --windowed \
     --hidden-import="zeroconf._utils.ipaddress" \
     --hidden-import="zeroconf._handlers.answers" \
     --add-data="src/onthespot/qt/qtui/*.ui:onthespot/qt/qtui" \
     --add-data="src/onthespot/resources/icons/*.png:onthespot/resources/icons" \
     --add-data="src/onthespot/resources/translations/*.qm:onthespot/resources/translations" \
-    --add-binary="build/ffmpeg:onthespot/bin/ffmpeg" \
+    $FFBIN \
     --paths="src/onthespot" \
     --name="OnTheSpot" \
     --icon="src/onthespot/resources/icons/onthespot.png" \
@@ -65,5 +77,6 @@ hdiutil create -srcfolder dist/dmg -format UDZO -o dist/OnTheSpot.dmg
 
 echo " => Cleaning up temporary files..."
 rm -rf __pycache__ build venv *.spec
+
 
 echo " => Done! .dmg available in 'dist/OnTheSpot.dmg'."
